feat: Concatenate algorithm implementations
